requirements.txt
flask==3.1.1
flask-cors==6.0.0
selenium==4.32.0
webdriver-manager==4.0.2
python-dotenv==1.1.0
gunicorn==21.2.0
requests==2.32.3 